% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_att.R
\name{aggregate_att}
\alias{aggregate_att}
\title{Title}
\usage{
aggregate_att(
  att_raw,
  type = c("simple", "dynamic", "calendar", "custom"),
  e = NULL
)
}
\arguments{
\item{att_raw}{Raw object as returned by DrDiD.}

\item{type}{Type of aggregation. "simple" returns the overall ATT, "dynamic"
returns event study type, "calendar" returns ATT aggregated by post-treatment
period, and "custom" returns aggregated as specified with argument "e".}

\item{e}{When type is "calendar", specify the range of event time to
aggregate. Returns a single value for this range.}
}
\value{
Summary table of estimated ATT.
}
\description{
Title
}
