% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Title}
\usage{
bootstrap(dp, B = 30)
}
\arguments{
\item{dp}{Pre-processed data object}

\item{B}{Number of bootstrap samples}
}
\value{
List of estimated group-time average treatment effects
}
\description{
Title
}
