% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRDID-package.R
\docType{package}
\name{DRDID-package}
\alias{DRDID}
\alias{DRDID-package}
\title{DRDID: Doubly Robust Difference-in-Differences Estimators}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements the locally efficient doubly robust difference-in-differences (DiD) estimators for the average treatment effect proposed by Sant'Anna and Zhao (2020) \doi{10.1016/j.jeconom.2020.06.003}. The estimator combines inverse probability weighting and outcome regression estimators (also implemented in the package) to form estimators with more attractive statistical properties. Two different estimation methods can be used to estimate the nuisance functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psantanna.com/DRDID/}
  \item \url{https://github.com/pedrohcgs/DRDID}
  \item Report bugs at \url{https://github.com/pedrohcgs/DRDID/issues}
}

}
\author{
\strong{Maintainer}: Pedro H. C. Sant'Anna \email{pedrosantanna@causal-solutions.com} [copyright holder]

Authors:
\itemize{
  \item Jun Zhao \email{beanzhaojun11@gmail.com}
}

}
\keyword{internal}
