% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drdid_imp_panel.R
\name{drdid_imp_panel}
\alias{drdid_imp_panel}
\title{Improved Doubly Robust Difference-in-Differences Estimator for the ATT, with Panel Data}
\usage{
drdid_imp_panel(
  y1,
  y0,
  D,
  covariates,
  i.weights = NULL,
  boot = F,
  boot.type = "weighted",
  nboot = NULL,
  inffunc = F
)
}
\arguments{
\item{y1}{An \eqn{n} x \eqn{1} vector of outcomes from the post-treatment period.}

\item{y0}{An \eqn{n} x \eqn{1} vector of outcomes from the pre-treatment period.}

\item{D}{An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).}

\item{covariates}{An \eqn{n} x \eqn{k} matrix of covariates to be used in the propensity score and regression estimation.
If covariates = NULL, this leads to an unconditional DID estimator.}

\item{i.weights}{An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.}

\item{boot}{Logical argument to whether bootstrap should be used for inference. Default is FALSE.}

\item{boot.type}{Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier".
If boot==T, default is "weighted".}

\item{nboot}{Number of bootstrap repetitions (not relevant if boot = FALSE). Default is 999 if boot = TRUE}

\item{inffunc}{Logical argument to whether influence function should be returned. Default is FALSE.}
}
\value{
A list containing the following components:
 \item{ATT}{The DID point estimate.}
 \item{se}{The DID standard error.}
 \item{uci}{The upper bound of the 95\% CI for the ATT.}
 \item{lci}{The lower bound of the 95\% CI for the ATT}
 \item{boots}{All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is \code{NULL}.}
 \item{ps.flag}{Convergence Flag for the propensity score estimation: =0 if \code{trust} algorithm converged,
   =1 if IPW algorithm converged (in case it was used), =2 if GLM logit estimator was used (i.e., if both \code{trust} and IPT
   did not converged).}
 \item{att.inf.func}{Estimate of the influence function. Default is NULL}
 \item{call.param}{The matched call.}
 \item{argu}{Some arguments used (explicitly or not) in the call (panel = T, estMethod = "imp", boot, boot.type, nboot, type="dr")}
}
\description{
Improved Doubly Robust Difference-in-Differences Estimator for the ATT, with Panel Data
}
\references{
{

\cite{Sant'Anna, Pedro H. C. and Zhao, Jun. (2020),
"Doubly Robust Difference-in-Differences Estimators." Journal of Econometrics, Forthcoming,
\url{https://arxiv.org/abs/1812.01723}}
}
}
