% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisance_models.R
\name{fit_nuisance_models}
\alias{fit_nuisance_models}
\title{Title}
\usage{
fit_nuisance_models(preprocessed_list)
}
\arguments{
\item{preprocessed_list}{List of pre-processed data object/list as returned
by "preprocess()".}
}
\value{
List of four models is returned: outcome model for control condition,
one outcome model for treatment condition for each pre- and post treatment,
and exposure/propensity model.
}
\description{
Title
}
