% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attgt.R
\name{get_attgt}
\alias{get_attgt}
\title{Title}
\usage{
get_attgt(mods, df)
}
\arguments{
\item{mods}{List of four nuisance models}

\item{df}{Pre-processed data}
}
\value{
A two-element object/list. "att" contains the group-time average
treatment effects in table. "df_long" is the table combines the single
group-time specific tables, i.e. treatment and control group pre and post
treatment with regard to the base period and specific treatment date of
a group-time.
}
\description{
Title
}
