% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw_did_panel.R
\name{ipw_did_panel}
\alias{ipw_did_panel}
\title{Inverse Probability Weighted Difference-in-Differences Estimator, with Panel Data}
\usage{
ipw_did_panel(
  y1,
  y0,
  D,
  covariates,
  i.weights = NULL,
  boot = F,
  boot.type = "weighted",
  nboot = NULL,
  inffunc = F
)
}
\arguments{
\item{y1}{An \eqn{n} x \eqn{1} vector of outcomes from the post-treatment period.}

\item{y0}{An \eqn{n} x \eqn{1} vector of outcomes from the pre-treatment period.}

\item{D}{An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).}

\item{covariates}{An \eqn{n} x \eqn{k} matrix of covariates to be used in the propensity score estimation.
If covariates = NULL, this leads to an unconditional DID estimator.}

\item{i.weights}{An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.}

\item{boot}{Logical argument to whether bootstrap should be used for inference. Default is FALSE.}

\item{boot.type}{Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier".
If boot==T, default is "weighted".}

\item{nboot}{Number of bootstrap repetitions (not relevant if boot = FALSE). Default is 999 if boot = TRUE.}

\item{inffunc}{Logical argument to whether influence function should be returned. Default is FALSE.}
}
\value{
A list containing the following components:
\item{ATT}{The IPW DID point estimate.}
\item{se}{ The IPW DID standard error}
\item{uci}{Estimate of the upper bound of a 95\% CI for the ATT}
\item{lci}{Estimate of the lower bound of a 95\% CI for the ATT}
\item{boots}{All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL}
\item{att.inf.func}{Estimate of the influence function. Default is NULL}
 \item{call.param}{The matched call.}
 \item{argu}{Some arguments used (explicitly or not) in the call (panel = T, normalized = F, boot, boot.type, nboot, type="ipw")}
}
\description{
Inverse Probability Weighted Difference-in-Differences Estimator, with Panel Data
}
\examples{
# -----------------------------------------------
# -----------------------------------------------
# Panel data in "wide" format
# -----------------------------------------------
# Data preparation using Lalonde sample with CPS comparison group
# Create "selection" treatment dummy: 1 if in experimental sample, 0 if in non-experimental
nsw$treated2 <- ifelse(is.na(nsw$treated), 0 , 1)

# Form the Lalonde sample with CPS comparison group
eval_lalonde_cps <- subset(nsw, nsw$treated == 0 | nsw$sample == 2)
# Select some covariates
covX = as.matrix(cbind(eval_lalonde_cps$age, eval_lalonde_cps$educ,
                             eval_lalonde_cps$black, eval_lalonde_cps$married,
                             eval_lalonde_cps$nodegree, eval_lalonde_cps$hisp,
                             eval_lalonde_cps$re74))
# -----------------------------------------------
# Implement (unnormalized) IPW DID with panel data
ipw_did_panel(y1 = eval_lalonde_cps$re78, y0 = eval_lalonde_cps$re75,
                D = eval_lalonde_cps$treated2,
                covariates = covX)

}
\references{
{
\cite{Abadie, Alberto (2005), "Semiparametric Difference-in-Differences Estimators",
Review of Economic Studies, vol. 72(1), p. 1-19, \url{https://doi.org/10.1111/0034-6527.00321}.
}

\cite{Sant'Anna, Pedro H. C. and Zhao, Jun. (2020),
"Doubly Robust Difference-in-Differences Estimators." Journal of Econometrics, Forthcoming,
\url{https://arxiv.org/abs/1812.01723}}


}
}
