% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_AIPW.R
\name{predict_AIPW}
\alias{predict_AIPW}
\title{Title}
\usage{
predict_AIPW(mods, df)
}
\arguments{
\item{mods}{List of four nuisance models}

\item{df}{(Long) pre-processed data}
}
\value{
Summary table of AIPW ATT estimates for each group-time in df.
}
\description{
Title
}
