% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_did_rc.R
\name{reg_did_rc}
\alias{reg_did_rc}
\title{Regression-based DID estimator with Repeated Cross Section Data
Regression-based Difference-in-Differences Estimator for the ATT, with Repeated Cross Section Data}
\usage{
reg_did_rc(
  y,
  post,
  D,
  covariates,
  i.weights = NULL,
  boot = F,
  boot.type = "weighted",
  nboot = NULL,
  inffunc = F
)
}
\arguments{
\item{y}{An \eqn{n} x \eqn{1} vector of outcomes from the both pre and post-treatment periods.}

\item{post}{An \eqn{n} x \eqn{1} vector of Post-Treatment dummies (post = 1 if observation belongs to post-tretment period,
and post = 0 if observation belongs to pre-treatment period.)}

\item{D}{An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).}

\item{covariates}{An \eqn{n} x \eqn{k} matrix of covariates to be used in the regression estimation.
If covariates = NULL, this leads to an unconditional DID estimator.}

\item{i.weights}{An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.}

\item{boot}{Logical argument to whether bootstrap should be used for inference. Default is FALSE.}

\item{boot.type}{Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier".
If boot==T, default is "weighted".}

\item{nboot}{Number of bootstrap repetitions (not relevant if boot = FALSE). Default is 999 if boot = TRUE.}

\item{inffunc}{Logical argument to whether influence function should be returned. Default is FALSE.}
}
\value{
A list containing the following components:
 \item{ATT}{The Reg DID point estimate}
 \item{se}{The Reg DID standard error}
 \item{uci}{Estimate of the upper boudary of a 95\% CI for the ATT}
 \item{lci}{Estimate of the lower boudary of a 95\% CI for the ATT}
 \item{boots}{All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL}
 \item{att.inf.func}{Estimate of the influence function. Default is NULL}
}
\description{
Regression-based DID estimator with Repeated Cross Section Data
Regression-based Difference-in-Differences Estimator for the ATT, with Repeated Cross Section Data
}
