% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\title{Title}
\usage{
statistics(att_0, att_b)
}
\arguments{
\item{att_0}{Main estimates of ATT}

\item{att_b}{Bootstrap estimates of ATT}
}
\value{
ATT along with bootstrap standard errors and more statistics.
}
\description{
Title
}
