% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_ipw_did_rc.R
\name{std_ipw_did_rc}
\alias{std_ipw_did_rc}
\title{Standardized Inverse probability weighted Difference-in-Differences Estimator with Repeated Cross Section Data}
\usage{
std_ipw_did_rc(y, post, D, covariates, i.weights = NULL, boot = F,
  boot.type = "weighted", nboot = NULL)
}
\arguments{
\item{y}{An \eqn{n} x \eqn{1} vector of outcomes from the both pre and post-treatment periods.}

\item{post}{An \eqn{n} x \eqn{1} vector of Post-Treatment dummies (post = 1 if observation belongs to post-tretment period,
and post = 0 if observation belongs to pre-treatment period.)}

\item{D}{An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).}

\item{covariates}{An \eqn{n} x \eqn{k} matrix of covariates to be used in the propensity score estimation}

\item{i.weights}{An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.}

\item{boot}{Logical argument to whether bootstrap should be used for inference. Deafault is FALSE.}

\item{boot.type}{Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier".
If boot==T, default is "weighted".}

\item{nboot}{Number of bootstrap repetitions (not relevant if boot = FALSE). Deafault is 999 if boot = TRUE.}
}
\value{
A list containing the following components:
\item{ATT}{The IPW DID point estimate.}
\item{se}{ The IPW DID standard error}
\item{uci}{Estimate of the upper boudary of a 95\% CI for the ATT}
\item{lci}{Estimate of the lower boudary of a 95\% CI for the ATT}
\item{boots}{All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL}
}
\description{
Standardized Inverse probability weighted Difference-in-Differences Estimator with Repeated Cross Section Data
}
