% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfe_did_panel.R
\name{twfe_did_panel}
\alias{twfe_did_panel}
\title{Two-Way Fixed Effects DID estimator with Panel Data
Two-Way Fixed Effects DID estimator with Panel Data}
\usage{
twfe_did_panel(y1, y0, D, covariates, i.weights = NULL, boot = F,
  boot.type = "weighted", nboot = NULL)
}
\arguments{
\item{y1}{An \eqn{n} x \eqn{1} vector of outcomes from the post-treatment period.}

\item{y0}{An \eqn{n} x \eqn{1} vector of outcomes from the pre-treatment period.}

\item{D}{An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).}

\item{covariates}{An \eqn{n} x \eqn{k} matrix of covariates to be used in the regression estimation}

\item{i.weights}{An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.}

\item{boot}{Logical argument to whether bootstrap should be used for inference. Deafault is FALSE.}

\item{boot.type}{Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier".
If boot==T, default is "weighted".}

\item{nboot}{Number of bootstrap repetitions (not relevant if boot = FALSE). Deafault is 999 if boot = TRUE.}
}
\value{
A list containing the following components:
 \item{ATT}{The TWFE DID point estimate}
 \item{se}{The TWFE DID standard error}
 \item{uci}{Estimate of the upper boudary of a 95\% CI for the TWFE parameter.}
 \item{lci}{Estimate of the lower boudary of a 95\% CI for the TWFE parameter.}
 \item{boots}{All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL}
}
\description{
Two-Way Fixed Effects DID estimator with Panel Data
Two-Way Fixed Effects DID estimator with Panel Data
}
