% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wols.ie.panel.R
\name{wols.ie.panel}
\alias{wols.ie.panel}
\title{Weighted OLS for Intrinsically Efficient DR-DID}
\usage{
wols.ie.panel(deltaY, D, int.cov, pscore, i.weights)
}
\arguments{
\item{deltaY}{An \eqn{n} x \eqn{1} vector of the difference of outcomes between post and pre treatment periods, i.e., Y_post - Y_pre.}

\item{D}{An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).}

\item{int.cov}{An \eqn{n} x \eqn{k} matrix of covariates to be used in the regression estimation (with intercept)}

\item{pscore}{An \eqn{n} x \eqn{1} vector of propensity score estimates.}

\item{i.weights}{An \eqn{n} x \eqn{1} vector of weights to be used.}
}
\value{
A list containing the following components:
\item{out.reg}{The fitted values of the outcome regression}
}
\description{
Compute Weighted OLS regression parameters for the (Intrinsically Efficient) Doubly-Robust DID estimator
}
\details{
This function implements the Weighted OLS estimator for the Not-Treated group that is used to construct the
(Intrinsically Efficient) Doubly Robust DID estimator, when Panel Data is available
}
\references{
Sant'Anna, Pedro H.C., and Zhao, Jun B. (2019) "Doubly Robust difference-in-difference estimators", \emph{Working Paper}.
}
