% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drdid.R
\name{drdid}
\alias{drdid}
\title{Doubly Robust Difference-in-Differences Estimators for the ATT}
\usage{
drdid(
  yname,
  tname,
  idname,
  dname,
  xformla = NULL,
  data,
  panel = TRUE,
  estMethod = c("imp", "trad"),
  weightsname = NULL,
  boot = FALSE,
  boot.type = c("weighted", "multiplier"),
  nboot = 999,
  inffunc = FALSE
)
}
\arguments{
\item{yname}{The name of the outcome variable.}

\item{tname}{The name of the column containing the time periods.}

\item{idname}{The name of the column containing the unit id name.}

\item{dname}{The name of the column containing the treatment group (=1 if observation is treated in the post-treatment, =0 otherwise)}

\item{xformla}{A formula for the covariates to include in the model. It should be of the form \code{~ X1 + X2}
(intercept should not be listed as it is always automatically included). Default is NULL which is equivalent to \code{xformla=~1}.}

\item{data}{The name of the data.frame that contains the data.}

\item{panel}{Whether or not the data is a panel dataset. The panel dataset should be provided in long format -- that
is, where each row corresponds to a unit observed at a particular point in time.  The default is TRUE.
Whens \code{panel=TRUE}, the variable \code{idname} must be set.  When \code{panel=FALSE}, the data is treated
as stationary repeated cross sections.}

\item{estMethod}{the method to estimate the nuisance parameters.
The default is "imp" which uses weighted least squares to estimate the outcome regressions and
inverse probability tilting to the estimate the the propensity score, leading to the improved locally efficient  DR DID estimator
proposed by Sant'Anna and Zhao (2020). The other alternative is "trad",
which then uses OLS to estimate outcome regressions and maximum likelihood to estimate propensity score. This leads
to the "traditional" locally efficient DR DID estimator proposed by Sant'Anna and Zhao (2020).}

\item{weightsname}{The name of the column containing the sampling weights. If NULL, then every observation has the same weights.}

\item{boot}{Logical argument to whether bootstrap should be used for inference. Default is \code{FALSE} and analytical
standard errors are reported.}

\item{boot.type}{Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier".
If \code{boot==TRUE}, default is "weighted".}

\item{nboot}{Number of bootstrap repetitions (not relevant if boot = \code{FALSE}). Default is 999.}

\item{inffunc}{Logical argument to whether influence function should be returned. Default is \code{FALSE}.}
}
\value{
A list containing the following components:
\item{ATT}{The DR DID point estimate}
\item{se}{ The DR DID standard error}
\item{uci}{Estimate of the upper bound of a 95\% CI for the ATT}
\item{lci}{Estimate of the lower bound of a 95\% CI for the ATT}
\item{boots}{All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL}
\item{att.inf.func}{Estimate of the influence function. Default is NULL}
 \item{ps.flag}{Convergence Flag for the propensity score estimation
 (only active if \code{estMethod = "imp"}.): =0 if \code{trust} algorithm converged,
   =1 if IPT (original) algorithm converged (in case it was used), =2 if GLM logit estimator was used (i.e., if both \code{trust} and IPT
   did not converged).}
\item{call.param}{The matched call.}
\item{argu}{Some arguments used in the call (panel, estMethod, boot, boot.type, nboot, type="dr")}
}
\description{
\code{drdid} computes the doubly robust estimators for the average treatment effect on the treated
 in DID setups. It can be used with panel or repeated cross section data. The nuisance parameters can be estimated
 using "traditional" methods (OLS for the outcome regressions and maximum likelihood for the propensity score), or
 "improved" methods (weighted least squares for the outcome regressions and
 inverse probability tilting for the propensity score). Estimators based on "improved methods" are also doubly
 robust for inference. See Sant'Anna and Zhao (2020) for a detailed description.
}
\references{
{

\cite{Sant'Anna, Pedro H. C. and Zhao, Jun. (2020),
"Doubly Robust Difference-in-Differences Estimators." Journal of Econometrics, Forthcoming,
\url{https://arxiv.org/abs/1812.01723}}
}
}
